requirements.txt
protobuf~=4.25.4
pillow==10.2.0
streamlit==1.31.1
streamlit-option-menu==0.3.12
google-generativeai==0.4.0